backend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.3.4
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.8.0
Django==5.2
django-cors-headers==4.3.0
django-filter==23.3
django-formtools==2.5.1
django-guardian==2.4.0
django-otp==1.2.2
django-phonenumber-field==6.4.0
django-two-factor-auth==1.15.3
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.0
gunicorn==21.2.0
iniconfig==2.1.0
kombu==5.5.3
packaging==25.0
Pillow==10.1.0
pluggy==1.5.0
prompt_toolkit==3.0.51
psutil==7.0.0
psycopg2-binary==2.9.9
PyJWT==2.10.1
pyotp==2.9.0
pypng==0.20220715.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-django==4.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
qrcode==7.4.2
redis==5.0.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.5.0
drf-yasg==1.21.7
